separate pipeline client into sync and async clients
