+++ deepset_cloud_sdk/__init__.py
@@ -5,8 +5,11 @@ import logging
 
 import structlog
 
-from deepset_cloud_sdk.workflows.pipeline_client import PipelineClient
-from deepset_cloud_sdk.workflows.pipeline_client.models import (
+from deepset_cloud_sdk._service.pipeline_service import (
+    DeepsetValidationError,
+    ErrorDetail,
+)
+from deepset_cloud_sdk.models import (
     BaseConfig,
     IndexConfig,
     IndexInputs,
@@ -16,10 +19,10 @@ from deepset_cloud_sdk.workflows.pipeline_client.models import (
     PipelineOutputs,
     PipelineOutputType,
 )
-from deepset_cloud_sdk.workflows.pipeline_client.pipeline_service import (
-    DeepsetValidationError,
-    ErrorDetail,
+from deepset_cloud_sdk.workflows.async_client.async_pipeline_client import (
+    AsyncPipelineClient,
 )
+from deepset_cloud_sdk.workflows.sync_client.pipeline_client import PipelineClient
 
 structlog.configure(
     wrapper_class=structlog.make_filtering_bound_logger(logging.INFO),
@@ -28,6 +31,7 @@ structlog.configure(
 log = structlog.get_logger()
 
 __all__ = [
+    "AsyncPipelineClient",
     "BaseConfig",
     "ErrorDetail",
     "PipelineClient",
