+++ deepset_cloud_sdk/workflows/__init__.py
@@ -1,6 +1,10 @@
 """Workflows for deepset AI platform SDK."""
 
-from deepset_cloud_sdk.workflows.pipeline_client.models import (
+from deepset_cloud_sdk._service.pipeline_service import (
+    DeepsetValidationError,
+    ErrorDetail,
+)
+from deepset_cloud_sdk.models import (
     BaseConfig,
     IndexConfig,
     IndexInputs,
@@ -10,13 +14,13 @@ from deepset_cloud_sdk.workflows.pipeline_client.models import (
     PipelineOutputs,
     PipelineOutputType,
 )
-from deepset_cloud_sdk.workflows.pipeline_client.pipeline_client import PipelineClient
-from deepset_cloud_sdk.workflows.pipeline_client.pipeline_service import (
-    DeepsetValidationError,
-    ErrorDetail,
+from deepset_cloud_sdk.workflows.async_client.async_pipeline_client import (
+    AsyncPipelineClient,
 )
+from deepset_cloud_sdk.workflows.sync_client.pipeline_client import PipelineClient
 
 __all__ = [
+    "AsyncPipelineClient",
     "BaseConfig",
     "ErrorDetail",
     "PipelineInputs",
