--- deepset_cloud_sdk/workflows/pipeline_client/__init__.py
@@ -1,31 +1,0 @@
-"""Package to enable importing pipelines and indexes to deepset AI platform."""
-
-from deepset_cloud_sdk.workflows.pipeline_client.models import (
-    BaseConfig,
-    IndexConfig,
-    IndexInputs,
-    IndexOutputs,
-    PipelineConfig,
-    PipelineInputs,
-    PipelineOutputs,
-    PipelineOutputType,
-)
-from deepset_cloud_sdk.workflows.pipeline_client.pipeline_client import PipelineClient
-from deepset_cloud_sdk.workflows.pipeline_client.pipeline_service import (
-    DeepsetValidationError,
-    ErrorDetail,
-)
-
-__all__ = [
-    "BaseConfig",
-    "DeepsetValidationError",
-    "ErrorDetail",
-    "PipelineClient",
-    "PipelineInputs",
-    "IndexInputs",
-    "IndexOutputs",
-    "PipelineOutputs",
-    "IndexConfig",
-    "PipelineConfig",
-    "PipelineOutputType",
-]
