+++ tests/unit/workflows/sync_client/test_pipeline_client.py
@@ -6,15 +6,15 @@ from haystack import AsyncPipeline, Pipeline
 
 from deepset_cloud_sdk._api.config import CommonConfig
 from deepset_cloud_sdk._api.deepset_cloud_api import DeepsetCloudAPI
-from deepset_cloud_sdk.workflows.pipeline_client import PipelineClient
-from deepset_cloud_sdk.workflows.pipeline_client.models import (
+from deepset_cloud_sdk._service.pipeline_service import PipelineService
+from deepset_cloud_sdk.models import (
     IndexConfig,
     IndexInputs,
     PipelineConfig,
     PipelineInputs,
     PipelineOutputs,
 )
-from deepset_cloud_sdk.workflows.pipeline_client.pipeline_service import PipelineService
+from deepset_cloud_sdk.workflows.sync_client.pipeline_client import PipelineClient
 
 
 class TestPipelineClientInit:
@@ -22,12 +22,12 @@ class TestPipelineClientInit:
 
     def test_init_with_env_vars_success(self, monkeypatch: pytest.MonkeyPatch) -> None:
         """Test successful client initialization with environment variables."""
-        monkeypatch.setattr("deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.API_KEY", "env-api-key")
+        monkeypatch.setattr("deepset_cloud_sdk.workflows.async_client.async_pipeline_client.API_KEY", "env-api-key")
         monkeypatch.setattr(
-            "deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.API_URL", "https://env-api-url.com"
+            "deepset_cloud_sdk.workflows.async_client.async_pipeline_client.API_URL", "https://env-api-url.com"
         )
         monkeypatch.setattr(
-            "deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.DEFAULT_WORKSPACE_NAME", "test-workspace"
+            "deepset_cloud_sdk.workflows.async_client.async_pipeline_client.DEFAULT_WORKSPACE_NAME", "test-workspace"
         )
         pc = PipelineClient()
 
@@ -49,47 +49,6 @@ class TestPipelineClientInit:
         assert pc._api_config.api_url == "https://test-api-url-explicit.com"
         assert pc._workspace_name == "test-workspace-explicit"
 
-    def test_init_with_mixed_config(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Test client initialization with partial explicit configuration."""
-        monkeypatch.setattr(
-            "deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.API_URL", "https://env-api-url.com"
-        )
-        monkeypatch.setattr(
-            "deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.DEFAULT_WORKSPACE_NAME", "test-workspace"
-        )
-
-        pc = PipelineClient(
-            api_key="custom-api-key",
-        )
-
-        assert isinstance(pc._api_config, CommonConfig)
-        assert pc._api_config.api_key == "custom-api-key"
-        assert pc._api_config.api_url == "https://env-api-url.com"  # From environment
-        assert pc._workspace_name == "test-workspace"  # from environment
-
-    def test_init_with_missing_api_key_raises_error(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        # Remove both environment variable and global variable that was loaded at import time
-        monkeypatch.delenv("API_KEY", raising=False)
-        monkeypatch.setattr("deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.API_KEY", "")
-
-        with pytest.raises(ValueError):
-            PipelineClient(
-                api_url="https://api.com", workspace_name="test-workspace"
-            )  # Empty API key should raise ValueError
-
-    def test_init_with_missing_api_url_uses_default_url(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        monkeypatch.setattr("deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.API_URL", "")
-
-        pc = PipelineClient(api_key="hello", api_url="", workspace_name="test-workspace")
-
-        assert pc._api_config.api_url == "https://api.cloud.deepset.ai/api/v1"
-
-    def test_init_with_missing_workspace_raises_error(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        monkeypatch.setattr("deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.DEFAULT_WORKSPACE_NAME", "")
-
-        with pytest.raises(ValueError):
-            PipelineClient(api_key="hello", api_url="https://api.com")  # Empty workspace name should raise ValueError
-
 
 class TestPipelineClientImport:
     """Test suite for importing Haystack pipelines into deepset."""
@@ -110,10 +69,10 @@ class TestPipelineClientImport:
         mock_pipeline_service = Mock(return_value=mock_service)
 
         monkeypatch.setattr(
-            "deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.DeepsetCloudAPI.factory", mock_api_factory
+            "deepset_cloud_sdk.workflows.async_client.async_pipeline_client.DeepsetCloudAPI.factory", mock_api_factory
         )
         monkeypatch.setattr(
-            "deepset_cloud_sdk.workflows.pipeline_client.pipeline_client.PipelineService", mock_pipeline_service
+            "deepset_cloud_sdk.workflows.async_client.async_pipeline_client.PipelineService", mock_pipeline_service
         )
 
         return {
@@ -144,46 +103,6 @@ class TestPipelineClientImport:
             inputs=PipelineInputs(query=["retriever.query"]),
             outputs=PipelineOutputs(answers="answer_builder.answers"),
         )
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize("pipeline_type", [Pipeline, AsyncPipeline])
-    async def test_import_into_deepset_async_and_index_config_success(
-        self,
-        pipeline_type: Pipeline | AsyncPipeline,
-        mock_api_service_setup: dict,
-        client_with_explicit_config: PipelineClient,
-        index_config: IndexConfig,
-    ) -> None:
-        """Test successful async import of pipeline with IndexConfig."""
-        mock_pipeline = Mock(spec=pipeline_type)
-
-        await client_with_explicit_config.import_into_deepset_async(mock_pipeline, index_config)
-
-        mock_api_service_setup["api_factory"].assert_called_once_with(client_with_explicit_config._api_config)
-        mock_api_service_setup["pipeline_service"].assert_called_once_with(
-            mock_api_service_setup["api"], "test-workspace"
-        )
-        mock_api_service_setup["service"].import_async.assert_called_once_with(mock_pipeline, index_config)
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize("pipeline_type", [Pipeline, AsyncPipeline])
-    async def test_import_into_deepset_async_and_pipeline_config_success(
-        self,
-        pipeline_type: Pipeline | AsyncPipeline,
-        mock_api_service_setup: dict,
-        client_with_explicit_config: PipelineClient,
-        pipeline_config: PipelineConfig,
-    ) -> None:
-        """Test successful async import of pipeline with PipelineConfig."""
-        mock_pipeline = Mock(spec=pipeline_type)
-
-        await client_with_explicit_config.import_into_deepset_async(mock_pipeline, pipeline_config)
-
-        mock_api_service_setup["api_factory"].assert_called_once_with(client_with_explicit_config._api_config)
-        mock_api_service_setup["pipeline_service"].assert_called_once_with(
-            mock_api_service_setup["api"], "test-workspace"
-        )
-        mock_api_service_setup["service"].import_async.assert_called_once_with(mock_pipeline, pipeline_config)
 
     @pytest.mark.parametrize("pipeline_type", [Pipeline, AsyncPipeline])
     def test_import_into_deepset_and_index_config_success(
